// src/app/models/incidencias.ts
export interface IncidenciasInterface {
    Aprovacion?: string;
    Checker?: string;
    Cine?: string;
    Descripcion?: string;
    Encabezado?: string;
    Fecha?: string;
    imageURL?: string;
    postID?: string;
    timeStamp?: string;
    key?: string;
  }



export interface IncidenciaInterface {
    Aprovacion?: string;
    Checker?: string;
    Cine?: string;
    Descripcion?: string;
    Encabezado?: string;
    Fecha?: string;
    imageURL?: string;
    postID?: string;
    timeStamp?: string;
    key?: string;
  }